+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/driver/FmuDriver.kt
@@ -24,7 +24,7 @@
 
 package no.ntnu.ihb.fmi4j.driver
 
-import no.ntnu.ihb.fmi4j.importer.Fmu
+import no.ntnu.ihb.fmi4j.importer.fmi2.Fmu
 
 
 /**
