+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi2/me/ModelExchangeInstance.kt
@@ -22,11 +22,11 @@
  * THE SOFTWARE.
  */
 
-package no.ntnu.ihb.fmi4j.importer.me
+package no.ntnu.ihb.fmi4j.importer.fmi2.me
 
 import no.ntnu.ihb.fmi4j.FmiStatus
-import no.ntnu.ihb.fmi4j.importer.AbstractFmuInstance
-import no.ntnu.ihb.fmi4j.importer.jni.EventInfo
+import no.ntnu.ihb.fmi4j.importer.fmi2.AbstractFmuInstance
+import no.ntnu.ihb.fmi4j.importer.fmi2.jni.EventInfo
 import no.ntnu.ihb.fmi4j.modeldescription.ModelExchangeModelDescription
 
 /**
