+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi2/me/ModelExchangeLibraryWrapper.kt
@@ -22,12 +22,12 @@
  * THE SOFTWARE.
  */
 
-package no.ntnu.ihb.fmi4j.importer.me
+package no.ntnu.ihb.fmi4j.importer.fmi2.me
 
 import no.ntnu.ihb.fmi4j.FmiStatus
-import no.ntnu.ihb.fmi4j.importer.jni.EventInfo
-import no.ntnu.ihb.fmi4j.importer.jni.Fmi2LibraryWrapper
-import no.ntnu.ihb.fmi4j.importer.jni.Fmi2ModelExchangeLibrary
+import no.ntnu.ihb.fmi4j.importer.fmi2.jni.EventInfo
+import no.ntnu.ihb.fmi4j.importer.fmi2.jni.Fmi2LibraryWrapper
+import no.ntnu.ihb.fmi4j.importer.fmi2.jni.Fmi2ModelExchangeLibrary
 import no.ntnu.ihb.fmi4j.util.BooleanByReference
 
 /**
