+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi2/me/ModelExchangeSlave.kt
@@ -22,7 +22,7 @@
  * THE SOFTWARE.
  */
 
-package no.ntnu.ihb.fmi4j.importer.me
+package no.ntnu.ihb.fmi4j.importer.fmi2.me
 
 import no.ntnu.ihb.fmi4j.FmiStatus
 import no.ntnu.ihb.fmi4j.FmuSlave
