+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/util/unzipper.kt
@@ -72,4 +72,3 @@ internal fun InputStream.extractContentTo(directory: File) {
     }
 
 }
-
