+++ fmi-import/src/test/java/no/ntnu/ihb/fmi4j/importer/ControlledTemperatureTestJava.java
@@ -27,6 +27,7 @@ package no.ntnu.ihb.fmi4j.importer;
 import no.ntnu.ihb.fmi4j.Fmi4jVariableUtils;
 import no.ntnu.ihb.fmi4j.FmiStatus;
 import no.ntnu.ihb.fmi4j.FmuSlave;
+import no.ntnu.ihb.fmi4j.importer.fmi2.CoSimulationFmu;
 import no.ntnu.ihb.fmi4j.modeldescription.variables.RealVariable;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
