+++ fmi-import/src/test/java/no/ntnu/ihb/fmi4j/importer/TestFMUs.java
@@ -1,5 +1,6 @@
 package no.ntnu.ihb.fmi4j.importer;
 
+import no.ntnu.ihb.fmi4j.importer.fmi2.Fmu;
 import no.ntnu.ihb.fmi4j.modeldescription.ModelDescriptionParser;
 import no.ntnu.ihb.fmi4j.modeldescription.ModelDescriptionProvider;
 import no.ntnu.ihb.fmi4j.modeldescription.fmi2.JaxbModelDescriptionParser;
