+++ fmi-import/src/test/kotlin/no/ntnu/ihb/fmi4j/importer/misc/TestIllegalFmuInstanceType.kt
@@ -1,6 +1,6 @@
 package no.ntnu.ihb.fmi4j.importer.misc
 
-import no.ntnu.ihb.fmi4j.importer.Fmu
+import no.ntnu.ihb.fmi4j.importer.fmi2.Fmu
 import no.ntnu.ihb.fmi4j.importer.TestFMUs
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
